renamed modal -> model
